+++ Sources/MapboxCoreNavigation/RouteController.swift
@@ -549,7 +549,7 @@ open class RouteController: NSObject {
     private func update(progress: RouteProgress, with location: CLLocation, rawLocation: CLLocation, upcomingRouteAlerts routeAlerts: [UpcomingRouteAlert], mapMatchingResult: MapMatchingResult, routeShapeIndex: Int) {
         progress.updateDistanceTraveled(with: location)
         progress.upcomingRouteAlerts = routeAlerts.map { RouteAlert($0) }
-        progress.currentRouteShapeIndex = routeShapeIndex
+        progress.shapeIndex = routeShapeIndex
         
         //Fire the delegate method
         delegate?.router(self, didUpdate: progress, with: location, rawLocation: rawLocation)